feat(contract): create ContractUtilityIncrease
Refs: #77

+++ backend/ms-users/src/main/java/pi/ms_users/dto/email/EmailContractIncreaseLoadedDTO.java
@@ -11,6 +11,6 @@ public class EmailContractIncreaseLoadedDTO {
     private String lastName;
     private BigDecimal newAmount;
     private String currency;
-    private BigDecimal increase;
+    private Integer increase;
     private String index;
 }
